+++ Editor/AutomaticConfiguration.cs
@@ -1,3 +1,4 @@
+using CustomLocalization4EditorExtension;
 using UnityEditor;
 using UnityEngine;
 
@@ -10,6 +11,8 @@ namespace Anatawa12.AvatarOptimizer
         private SerializedProperty _removeUnusedObjects;
         private SerializedProperty _mmdWorldCompatibility;
         private SerializedProperty _advancedAnimatorParser;
+        private SerializedProperty _advancedSettings;
+        private GUIContent _advancedSettingsLabel = new GUIContent();
 
         private void OnEnable()
         {
@@ -17,6 +20,7 @@ namespace Anatawa12.AvatarOptimizer
             _removeUnusedObjects = serializedObject.FindProperty(nameof(TraceAndOptimize.removeUnusedObjects));
             _mmdWorldCompatibility = serializedObject.FindProperty(nameof(TraceAndOptimize.mmdWorldCompatibility));
             _advancedAnimatorParser = serializedObject.FindProperty(nameof(TraceAndOptimize.advancedAnimatorParser));
+            _advancedSettings = serializedObject.FindProperty(nameof(TraceAndOptimize.advancedSettings));
         }
 
         protected override void OnInspectorGUIInner()
@@ -26,10 +30,25 @@ namespace Anatawa12.AvatarOptimizer
 
             GUILayout.Label("General Settings", EditorStyles.boldLabel);
             EditorGUILayout.PropertyField(_mmdWorldCompatibility);
-            EditorGUILayout.PropertyField(_advancedAnimatorParser);
             GUILayout.Label("Features", EditorStyles.boldLabel);
             EditorGUILayout.PropertyField(_freezeBlendShape);
             EditorGUILayout.PropertyField(_removeUnusedObjects);
+
+            _advancedSettingsLabel.text = CL4EE.Tr("TraceAndOptimize:prop:advancedSettings");
+            if (EditorGUILayout.PropertyField(_advancedSettings, _advancedSettingsLabel, false))
+            {
+                EditorGUI.indentLevel++;
+                EditorGUILayout.HelpBox(CL4EE.Tr("TraceAndOptimize:warn:advancedSettings"), MessageType.Warning);
+                EditorGUILayout.PropertyField(_advancedAnimatorParser);
+                var iterator = _advancedSettings.Copy();
+                var enterChildren = true;
+                while (iterator.NextVisible(enterChildren))
+                {
+                    enterChildren = false;
+                    EditorGUILayout.PropertyField(iterator);
+                }
+                EditorGUI.indentLevel--;
+            }
 
             serializedObject.ApplyModifiedProperties();
         }